+++ lib/zoi/types/decimal.ex
@@ -2,7 +2,7 @@ if Code.ensure_loaded?(Decimal) do
   defmodule Zoi.Types.Decimal do
     @moduledoc false
 
-    use Zoi.Type.Def, fields: [:gte, :lte, :gt, :lt, coerce: false]
+    use Zoi.Type.Def, fields: [:gte, :lte, :gt, :lt, :multiple_of, coerce: false]
 
     alias Zoi.Validations
 
@@ -31,12 +31,17 @@ if Code.ensure_loaded?(Decimal) do
           Zoi.Opts.constraint_schema(Zoi.Types.Decimal.new([]),
             description: "decimal less than",
             error: error
+          ),
+        multiple_of:
+          Zoi.Opts.constraint_schema(Zoi.Types.Decimal.new([]),
+            description: "decimal must be multiple of",
+            error: error
           )
       )
     end
 
     def new(opts \\ []) do
-      {validation_opts, opts} = Keyword.split(opts, [:gte, :lte, :gt, :lt])
+      {validation_opts, opts} = Keyword.split(opts, [:gte, :lte, :gt, :lt, :multiple_of])
 
       opts
       |> apply_type()
@@ -44,6 +49,7 @@ if Code.ensure_loaded?(Decimal) do
       |> Validations.maybe_set_validation(Validations.Lte, validation_opts[:lte])
       |> Validations.maybe_set_validation(Validations.Gt, validation_opts[:gt])
       |> Validations.maybe_set_validation(Validations.Lt, validation_opts[:lt])
+      |> Validations.maybe_set_validation(Validations.MultipleOf, validation_opts[:multiple_of])
     end
 
     defimpl Zoi.Type do
@@ -74,7 +80,8 @@ if Code.ensure_loaded?(Decimal) do
           {Validations.Gte, schema.gte},
           {Validations.Lte, schema.lte},
           {Validations.Gt, schema.gt},
-          {Validations.Lt, schema.lt}
+          {Validations.Lt, schema.lt},
+          {Validations.MultipleOf, schema.multiple_of}
         ]
         |> Validations.run_validations(schema, input)
       end
@@ -146,6 +153,22 @@ if Code.ensure_loaded?(Decimal) do
       end
     end
 
+    defimpl Zoi.Validations.MultipleOf do
+      def set(schema, value, opts \\ []) do
+        %{schema | multiple_of: {value, opts}}
+      end
+
+      def validate(_schema, input, value, opts) do
+        remainder = Decimal.rem(input, value)
+
+        if Decimal.eq?(remainder, Decimal.new(0)) do
+          :ok
+        else
+          {:error, Zoi.Error.not_multiple_of(value, opts)}
+        end
+      end
+    end
+
     defimpl Inspect do
       alias Zoi.Validations
 
@@ -154,7 +177,8 @@ if Code.ensure_loaded?(Decimal) do
           gte: Validations.unwrap_validation(type.gte),
           lte: Validations.unwrap_validation(type.lte),
           gt: Validations.unwrap_validation(type.gt),
-          lt: Validations.unwrap_validation(type.lt)
+          lt: Validations.unwrap_validation(type.lt),
+          multiple_of: Validations.unwrap_validation(type.multiple_of)
         ]
 
         Zoi.Inspect.build(type, opts, extra_fields)
@@ -168,6 +192,7 @@ if Code.ensure_loaded?(Decimal) do
         |> maybe_add(:maximum, schema.lte)
         |> maybe_add(:exclusiveMinimum, schema.gt)
         |> maybe_add(:exclusiveMaximum, schema.lt)
+        |> maybe_add(:multipleOf, schema.multiple_of)
       end
 
       defp maybe_add(map, _key, nil), do: map