+++ lib/zoi/types/integer.ex
@@ -1,7 +1,7 @@
 defmodule Zoi.Types.Integer do
   @moduledoc false
 
-  use Zoi.Type.Def, fields: [:gte, :lte, :gt, :lt, coerce: false]
+  use Zoi.Type.Def, fields: [:gte, :lte, :gt, :lt, :multiple_of, coerce: false]
 
   alias Zoi.Validations
 
@@ -30,12 +30,17 @@ defmodule Zoi.Types.Integer do
         Zoi.Opts.constraint_schema(Zoi.Types.Integer.new([]),
           description: "integer less than",
           error: error
+        ),
+      multiple_of:
+        Zoi.Opts.constraint_schema(Zoi.Types.Integer.new([]),
+          description: "integer must be multiple of",
+          error: error
         )
     )
   end
 
   def new(opts \\ []) do
-    {validation_opts, opts} = Keyword.split(opts, [:gte, :lte, :gt, :lt])
+    {validation_opts, opts} = Keyword.split(opts, [:gte, :lte, :gt, :lt, :multiple_of])
 
     opts
     |> apply_type()
@@ -43,6 +48,7 @@ defmodule Zoi.Types.Integer do
     |> Validations.maybe_set_validation(Validations.Lte, validation_opts[:lte])
     |> Validations.maybe_set_validation(Validations.Gt, validation_opts[:gt])
     |> Validations.maybe_set_validation(Validations.Lt, validation_opts[:lt])
+    |> Validations.maybe_set_validation(Validations.MultipleOf, validation_opts[:multiple_of])
   end
 
   defimpl Zoi.Type do
@@ -71,7 +77,8 @@ defmodule Zoi.Types.Integer do
         {Validations.Gte, schema.gte},
         {Validations.Lte, schema.lte},
         {Validations.Gt, schema.gt},
-        {Validations.Lt, schema.lt}
+        {Validations.Lt, schema.lt},
+        {Validations.MultipleOf, schema.multiple_of}
       ]
       |> Validations.run_validations(schema, input)
     end
@@ -143,6 +150,20 @@ defmodule Zoi.Types.Integer do
     end
   end
 
+  defimpl Zoi.Validations.MultipleOf do
+    def set(schema, value, opts \\ []) do
+      %{schema | multiple_of: {value, opts}}
+    end
+
+    def validate(_schema, input, value, opts) do
+      if rem(input, value) == 0 do
+        :ok
+      else
+        {:error, Zoi.Error.not_multiple_of(value, opts)}
+      end
+    end
+  end
+
   defimpl Inspect do
     alias Zoi.Validations
 
@@ -151,7 +172,8 @@ defmodule Zoi.Types.Integer do
         gte: Validations.unwrap_validation(type.gte),
         lte: Validations.unwrap_validation(type.lte),
         gt: Validations.unwrap_validation(type.gt),
-        lt: Validations.unwrap_validation(type.lt)
+        lt: Validations.unwrap_validation(type.lt),
+        multiple_of: Validations.unwrap_validation(type.multiple_of)
       ]
 
       Zoi.Inspect.build(type, opts, extra_fields)
@@ -165,6 +187,7 @@ defmodule Zoi.Types.Integer do
       |> maybe_add(:maximum, schema.lte)
       |> maybe_add(:exclusiveMinimum, schema.gt)
       |> maybe_add(:exclusiveMaximum, schema.lt)
+      |> maybe_add(:multipleOf, schema.multiple_of)
     end
 
     defp maybe_add(map, _key, nil), do: map